refactor(storage): extract sqlite_store cache layer to _history_caches
A2 of the sqlite_store decomposition. Moves the three cache tables'
helpers (19 methods across run_context_cache, column_comments_cache,
schemas_cache) out of the 3512-LOC SQLiteHistoryStore into a focused
amx/storage/_history_caches.py module.

Each function takes the store as 'hs' and reuses its _lock +
_connect() plumbing. No DDL touched - all CREATE TABLE / ALTER TABLE
statements in init() stay byte-identical. The user's local history.db
file is unaffected.

SQLiteHistoryStore keeps thin delegating wrappers (passing
*args/**kwargs so we preserve every signature unchanged) on every
name so the dozens of callers (orchestrator, connector, /db cache
CLI commands, web routers, test mocks) stay byte-compatible.

sqlite_store.py: 3512 -> 2924 LOC (-588, -17%).
